# Update gather to use multiple threads