--- a/onnxruntime/test/providers/cpu/tensor/gather_elements_op_test.cc
+++ b/onnxruntime/test/providers/cpu/tensor/gather_elements_op_test.cc
@@ -90,13 +90,7 @@ void RunTypedTest() {
   test5.AddOutput<T>("output", {2, 2},
                      {1, 1,
                       3, 3});
-  // skip nuphar, which will not throw error message but will ensure no out-of-bound access
-  // skip cuda as the cuda kernel won't throw the error message
-  // skip openvino which will not throw error message but will ensure no out-of-bound access
-  // skip TensorRT because it doesn't support out of bounds indices
-  test5.Run(OpTester::ExpectResult::kExpectFailure,
-            "GatherElements op: Value in indices must be within bounds [-2 , 1]. Actual value is 2",
-            {kNupharExecutionProvider, kCudaExecutionProvider, kRocmExecutionProvider, kOpenVINOExecutionProvider, kTensorrtExecutionProvider});
+  test5.Run(OpTester::ExpectResult::kExpectSuccess, "", {kTensorrtExecutionProvider});
 
   // 3D input - axis 1
   OpTester test6("GatherElements", 11);
@@ -247,11 +241,7 @@ void RunTypedTest<std::string>() {
   test4.AddOutput<std::string>("output", {2, 2},
                                {"a", "a",
                                 "c", "c"});
-  // skip nuphar, which will not throw error message but will ensure no out-of-bound access
-  // skip Openvino, which will not throw error message but will ensure no out-of-bound access
-  test4.Run(OpTester::ExpectResult::kExpectFailure,
-            "GatherElements op: Value in indices must be within bounds [-2 , 1]. Actual value is -3",
-            {kNupharExecutionProvider, kOpenVINOExecutionProvider});
+  test4.Run(OpTester::ExpectResult::kExpectSuccess, "", {kTensorrtExecutionProvider});
 
   // 3D input - axis 1
   OpTester test5("GatherElements", 11);