--- a/onnxruntime/test/providers/cpu/tensor/gather_elements_op_test.cc
+++ b/onnxruntime/test/providers/cpu/tensor/gather_elements_op_test.cc
@@ -78,7 +78,6 @@ void RunTypedTest() {
   // skip TensorRT because it doesn't support negative indices				  
   test4.Run(OpTester::ExpectResult::kExpectSuccess, "", {kTensorrtExecutionProvider});
 
-#if 0
   // indices out of bounds
   OpTester test5("GatherElements", 11);
   test5.AddAttribute<int64_t>("axis", 1LL);
@@ -91,8 +90,13 @@ void RunTypedTest() {
   test5.AddOutput<T>("output", {2, 2},
                      {1, 1,
                       3, 3});
-  test5.Run(OpTester::ExpectResult::kExpectSuccess, "", {kTensorrtExecutionProvider});
-#endif
+  // skip nuphar, which will not throw error message but will ensure no out-of-bound access
+  // skip cuda as the cuda kernel won't throw the error message
+  // skip openvino which will not throw error message but will ensure no out-of-bound access
+  // skip TensorRT because it doesn't support out of bounds indices
+  test5.Run(OpTester::ExpectResult::kExpectFailure,
+            "GatherElements op: Out of range value in index tensor",
+            {kNupharExecutionProvider, kCudaExecutionProvider, kRocmExecutionProvider, kOpenVINOExecutionProvider, kTensorrtExecutionProvider});
 
   // 3D input - axis 1
   OpTester test6("GatherElements", 11);
@@ -231,7 +235,6 @@ void RunTypedTest<std::string>() {
                                 "d", "d"});
   test3.Run();
 
-#if 0
   // indices out of bounds
   OpTester test4("GatherElements", 11);
   test4.AddAttribute<int64_t>("axis", 1LL);
@@ -244,8 +247,11 @@ void RunTypedTest<std::string>() {
   test4.AddOutput<std::string>("output", {2, 2},
                                {"a", "a",
                                 "c", "c"});
-  test4.Run(OpTester::ExpectResult::kExpectSuccess, "", {kTensorrtExecutionProvider});
-#endif
+  // skip nuphar, which will not throw error message but will ensure no out-of-bound access
+  // skip Openvino, which will not throw error message but will ensure no out-of-bound access
+  test4.Run(OpTester::ExpectResult::kExpectFailure,
+            "GatherElements op: Out of range value in index tensor",
+            {kNupharExecutionProvider, kOpenVINOExecutionProvider});
 
   // 3D input - axis 1
   OpTester test5("GatherElements", 11);