**Description**: GatherElements wouldn't distribute the work across multiple threads

**Motivation and Context**
A user was comparing the performance of Onnxruntime vs Pytorch and saw that the latest Pytorch was 2x faster than Onnxruntime. A profile showed that they were using all CPU cores but we were limited to 1.

There is a separate issue where they were using Onnxruntime inefficiently (there's a memcpy on every Run() call to copy the output tensor, using io-bindings avoids the memcpy).

Here's some performance data comparing the old vs new. Note that there is a slight perf hit for the single threaded case, as it has to divide up the work into independent chunks vs a slightly faster incremental calculation between chunks.

New version using 8 threads:

onnx model: 0.770249s after 10000 iterations
pytorch model: 3.669060s after 10000 iterations

New version limited to one thread:

onnx model: 3.474726s after 10000 iterations
pytorch model: 3.579055s after 10000 iterations

Old version:

onnx model: 2.905542s after 10000 iterations
pytorch model: 3.634197s after 10000 iterations